fixing mistakes

+++ Data-definition-Managing-Tables-And-Working-With-Data-Types/time-to-practice-problem/02-create-products-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE products(
     product_name VARCHAR(300),
-    price INT,
+    price NUMERIC(10, 2),
     description TEXT,
     amount_in_stock INT,
     image VARCHAR(300)
